# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CI/CD Practice API
  version: 1.0.0
  description: Practice FastAPI backend for CI/CD pipeline learning.
servers:
  - url: http://localhost:8000
    description: Local / CI runner
paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      tags:
        - system
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - environment
                  - version
                properties:
                  status:
                    type: string
                    example: ok
                  environment:
                    type: string
                    example: test
                  version:
                    type: string
                    example: abc123
  /api/v1/items:
    get:
      summary: List items
      operationId: listItems
      tags:
        - items
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
    post:
      summary: Create item
      operationId: createItem
      tags:
        - items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemCreate"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /api/v1/items/{item_id}:
    get:
      summary: Get item by ID
      operationId: getItem
      tags:
        - items
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Item found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Item not found
components:
  schemas:
    ItemCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
    Item:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true